requirements.txt
streamlit
pandas
numpy
google-generativeai
twilio
python-dotenv
libsql-client
flask